+++ src/main/java/codeemoji/inlay/structuralanalysis/element/method/ExternalFunctionalityInvokingMethodSettings.java
@@ -1,6 +1,5 @@
 package codeemoji.inlay.structuralanalysis.element.method;
 
-import codeemoji.core.config.CEPSIType;
 import codeemoji.core.settings.CEBaseSettings;
 import com.intellij.openapi.components.State;
 import com.intellij.openapi.components.Storage;
@@ -17,7 +16,7 @@ public class ExternalFunctionalityInvokingMethodSettings extends CEBaseSettings<
     private boolean checkMethodCallsForExternalityApplied = false;
 
     public ExternalFunctionalityInvokingMethodSettings(){
-        super(builder().targetMethods().targetReferences(),
+        super(builder().targetMethods().targetReferences().targetsExternal(),
                 ExternalFunctionalityInvokingMethod.class, EXTERNAL_FUNCTIONALITY_INVOKING_METHOD);
     }